telegrambot/requirements-test.txt
Flask>=1.1.1
python-telegram-bot>=12.3.0
google-cloud-firestore>=1.6.1
python-dotenv>=0.10.5